Regime labeling
-Labeling regimes by grouping based on mean close_pc did not work well to accurately label predicted regimes
-Switched to trying every regime label permutation and using the best performing labeling order
-Added plotting of fills like sentiment and regime prediction in plot.py instead of a useless line
-Replaced ema_k and ema_d with ema_{period} with ma_periods in settings (renamed from sma_periods)
-Started integreating regime prediction into RNN training

+++ finbert_news.py
@@ -79,7 +79,7 @@ class FinBERTNews(object):
 
                 with torch.amp.autocast(self.device):  # Enable mixed precision
                     sentiment_probs = self.model(tokens["input_ids"], attention_mask=tokens["attention_mask"])["logits"]
-                    sentiment_probs = torch.nn.functional.softmax(torch.sum(sentiment_probs, 0), dim=-1).numpy()
+                    sentiment_probs = torch.nn.functional.softmax(torch.sum(sentiment_probs, 0), dim=-1).to(torch.float64).numpy()
                     sentiment = sentiment_probs[0] - sentiment_probs[1]  # positive% - negative%
         else:
             with torch.no_grad():  # Don't need gradients since we aren't training
